Removed dependence on MSBuild Community Tasks.
I had added an import of MSBuild.Community.Tasks.Targets to the Mercent.SqlServer.Management.csproj back on 4/20/2006 to support a Merge MSBuild target that merged some dependent .dlls with the .exe (I think there was a COM wrapper). I removed the Merge target back when I updated the project to work with SQL 2008 (and the dependent .dll was no longer necessary). At the time I didn't remember to remove the import of MSBuild Community Tasks.