Include extended properties in scripts.
All we really care about is the description ('MS_Description') extended property so that we can use database documentation tools (e.g. http://www.livedoco.com/). If our scripts begin to get cluttered with non-documentation extended properties we may want to consider how to only include the description extended properties.

I added the ExtendedProperties option to the ScriptingOptions objects used to script things out. I set the property for all types of database objects even though some of these objects may not even support extended properties.

I also ran into an issue when scripting tables since tables are actually scripted as multiple files. The extended properties of the table and columns were included in each file (the additional files are for keys, constraints, indexes, and foreign keys). To avoid these duplicates I added a HashSet that keeps track of the unique "EXEC sys.sp_addextendedproperty" and then skip writing it to the file if it is already in the set.