Fixed bcp file issue for tables with calculated or timestamp columns.
The previous changes that omitted calculated and timestamp columns from the bcp data files worked to script out the data files. But it did not work when loading the bcp data when creating the database from scripts.

Bcp with an XML format file requires the data file to include all columns (or else load into a view or use `INSERT ... SELECT ... FROM OPENROWSET ...`).
I still don't want calculated and timestamp (in particular) columns to be part of the scripts because I don't want differences in the values of these columns to result in a change that needs to be committed to source control. More importantly, I don't want it to affect the file-compare based verification used by the upgrade script generation process.