Fixed dependency order issues with data upgrade script.
I had scripted updates for all tables, in dependency order, then scripted inserts for all tables, in dependency order. However, that did not work when inserting a new row in table A and updating an existing row in table B to refer to the new row in table A. With this approach the update to table B was scripted before the insert to table A (all updates were scripted before all inserts). This would fail due to a foreign key violation.

Now we loop through the tables in dependency order and for each table script updates then inserts. With this approach the insert to table A is scripted before the update to table B.

To verify that this approach should always work, I made the equivalent of a truth table, with values I, U, and D (insert, update, delete) instead of T and F (true and false) for tables A and B and determined which table needed to be first.