Added ability to sync a target database.
This is similar to generating an upgrade script except the intent is not to generate a script that will later be used to update a production database. Instead, the purpose is to sync a target database now.

In sync mode the target database may have lots of data that is not part of the database creating scripts (it is not "static" data) is is therefore not in the source database. Using the upgrade process this data would be deleted. But using sync this data will be preserved. Sppecifically, tables have rows in the target but are empty in the source will not be deleted.

Other changes:
 * Added options to run without prompting (-f and -n) so the tool could be used by an automation script with no user interaction.
 * Added DataUpgradeReport.xml and show tables with deletes in console output.