Changed handling of file creation time in *.cab data files.
It appears that MakeCab.exe uses the current offset in some date/time calculations rather than using the offset appropriate for the file creation date (i.e. is daylight savings in effect now vs then). As a result, if ScriptDB is run in Pacific Daylight Time then it will generate a *.cab file with a couple bytes different than the equivalent file that ScriptDB generated when run in Pacific Standard Time.