Add delay to reduce chance of intermittent exceptions.
We have experienced some intermittent exceptions at the end up the process for generating upgrade scripts when we generate create scripts concurrently. It surprises me, but there appears to be some threading issues in the SMO library even though we are using separate instances of the Server and Database classes. Most of the issues occur during prefetch. We can minimize the frequency of these issues by waiting for 10 seconds before starting to generate scripts for the target database.